Add tests for classes and split them to separated files
