+++ src/test/calendar.spec.ts
@@ -1,26 +1,5 @@
-import Tooltip from '../assets/ts/lib/Calendar/tooltip';
 import Calendar from '../assets/ts/lib/Calendar/calendar';
 import moment from 'moment';
-
-describe('Tooltip class', () => {
-    describe('Create tooltip for parent elemenent', () => {
-        const message = "Hello world!";
-        const parentElement = document.createElement('div');
-        const tooltip = new Tooltip(parentElement, message);
-
-        it('should append tooltip with "Hello world!" message to parentElement', () => {
-            expect(tooltip.message).toBe("Hello world!")
-            expect(parentElement.children.length).toEqual(1);
-            expect(parentElement.innerHTML).toEqual('<div class="tip">Hello world!</div>');
-        });
-        it('should change message "Hello world!" in tooltip to "Goodbye"', () => {
-            tooltip.message = "Goodbye";
-            expect(tooltip.message).not.toBe("Hello world!");
-            expect(tooltip.message).toBe("Goodbye");
-        });
-    });
-});
-
 
 describe('Calendar class', () => {
     document.body.innerHTML = '<div id="calendar"></div>';
@@ -44,5 +23,3 @@ describe('Calendar class', () => {
         expect(nav.children[2].innerHTML.toLocaleLowerCase().trim()).toEqual('2018')
     });
 })
-
-
